obj: fix recycler not locating unused chunks
In scenarios where free() is done very rarely the memory in the recycler
would not be looked at because the estimates would indicate that no memory
was freed and there's no reason to search the tree. Normally that's OK,
because eventually the heap would trigger a forced search of the entire
recycler for the allocation class, locating all the memory that might have
been missed. But, the forced recalculation erronously looked at the
estimates and required that at least some memory were freed since the
last recycler run - which might have not been the case.
This patch makes the forced recycling run without looking at the estimates.

Reported-by: Denny Zhao <denny.zhao@memverge.com>